pyproject.toml
[project]
name = "licensemanager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.2",
    "djangorestframework>=3.15.2",
    "drf-yasg>=1.21.9",
]
